Make the module shaper::run_handler public and add an external test which proves that implementing a RunHandler is possible
